fix(growth): AutomatedEmailLog user_id should be INT not BIGINT (User.id is INT)
- Prisma schema: user_id Int (was BigInt)
- SQL migration: SERIAL/INTEGER (was BIGSERIAL/BIGINT)
- Route: removed all BigInt() wrappers
- Production DB: table dropped and recreated with INT columns
- Keepalive: use absolute URL for fetch (request.url may not have host)

This was causing the Vercel build to fail — Prisma couldn't generate
the client due to foreign key type mismatch.

+++ prisma/sql/2026-08-28-automated-email-sequence.sql
@@ -237,8 +237,8 @@ VALUES
 -- One row per (user_id, sequence, step) — UNIQUE constraint prevents re-sends.
 
 CREATE TABLE IF NOT EXISTS "AutomatedEmailLog" (
-  id BIGSERIAL PRIMARY KEY,
-  user_id BIGINT NOT NULL REFERENCES "User"(id) ON DELETE CASCADE,
+  id SERIAL PRIMARY KEY,
+  user_id INTEGER NOT NULL REFERENCES "User"(id) ON DELETE CASCADE,
   sequence TEXT NOT NULL,           -- 'welcome', 'profile_nudge', 'reengage'
   step TEXT NOT NULL,                -- 'day0', 'day1', 'day3', 'day7', 'day14', 'nudge_25', etc.
   template_key TEXT NOT NULL,        -- 'welcome_day0', 'profile_nudge_25', etc.
